Create mainframe Class
Create a special file to put the graphical interface.
The code is more modular.
